# handler/delete_opening.go
package handler

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/guimassoqueto/go-jobs/schemas"
)

// @BasePath /api/v1
// @Summary Delete opening
// @Description Delete a new Job Opening
// @Accept json
// @Produce json
// @Param id query string true "Opening identification"
// @Success 200 {object} ShowOpeningResponse
// @Failure 400 {object} ErrorResponse
// @Failure 404 {object} ErrorResponse
// @Router /opening [delete]
func DeleteOpeningHandler(ctx *gin.Context) {
	id := ctx.Query("id")
	if id == "" {
		sendError(ctx, http.StatusBadRequest, errParamIsRequired("id", "queryParameter").Error())
		return
	}

	opening := schemas.Opening{}
	if err := db.First(&opening, id).Error; err != nil {
		sendError(ctx, http.StatusNotFound, fmt.Sprintf("opening with id: %s not found", id))
		return 
	}

	if err := db.Delete(&opening).Error; err != nil {
		sendError(ctx, http.StatusNotFound, fmt.Sprintf("error deleting id: %s", id))
		return 
	}
	sendSuccess(ctx, "OK", opening)
}